+++ library/src/main/java/com/drunkenboys/ckscalendar/monthcalendar/MonthPageAdapter.kt
@@ -1,4 +1,4 @@
-package com.drunkenboys.ckscalendar.month
+package com.drunkenboys.ckscalendar.monthcalendar
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
